refactor: types are now lifted up to root directory
