+++ .gitignore
@@ -16,6 +16,7 @@ cypress.env.json
 # next.js
 /.next/
 /out/
+*.tsbuildinfo
 
 # production
 /build
